test/spark_server/pyproject.toml
[project]
name = "spark-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyspark-connect==4.1.1",
]
